requirements.txt
Pillow
pdf2image
pytesseract